requirements.txt
fastapi
uvicorn
python-dotenv
requests
psycopg2-binary
gunicorn
python-multipart
jinja2
